Added mentor page Modified the helper function which redirects after login

+++ app/helpers/application_helper.rb
@@ -5,6 +5,10 @@ module ApplicationHelper
   end
 
   def return_from_signin
-    redirect_to session.delete(:return_to)
+    if session[:return_to]
+      redirect_to session.delete(:return_to)
+    else
+      redirect_to root_path
+    end
   end
 end
